requirements.txt
django
google.generativeai
pillow
together
supabase
pexels-api
